made the header/navigation follow the window for ease of access. and adjustments to fonts and padding using utopia
the main effort was fixing the position: fixed for the header (which took a hell lot of time and time to diagnose what was causing the position: fixed to behave as position: absolute). so now the navigation follows the window for ease of access (used js for some cool animation). also, some adjustments to the fonts and padding (more to come) were made to improve the responsiveness (using utopia). and some miscellaneous adjustment